+++ packages/auto_animated/lib/src/sliver_grid.dart
@@ -10,22 +10,15 @@ const Duration _kDuration = Duration(milliseconds: 300);
 class AutoAnimatedSliverGrid extends StatefulWidget {
   ///  when they are inserted or removed.
   const AutoAnimatedSliverGrid({
-    @required
-        this.itemBuilder,
-    @required
-        this.gridDelegate,
-    Key key,
+    @required this.itemBuilder,
+    @required this.gridDelegate,
+    @required this.itemCount,
     this.delay = Duration.zero,
     this.reverse = false,
     this.showItemInterval = _kDuration,
     this.showItemDuration = _kDuration,
-    @Deprecated('Usage `itemCount` instead '
-        '(without character "s"). Will be deleted in 1.2.0')
-        int itemsCount = 0,
-    int itemCount = 0,
+    Key key,
   })  : assert(itemBuilder != null),
-        // ignore: deprecated_member_use_from_same_package
-        itemCount = itemCount ?? itemsCount,
         assert(itemCount != null && itemCount >= 0),
         super(key: key);
 
@@ -108,7 +101,7 @@ class _AutoAnimatedSliverGridState extends State<AutoAnimatedSliverGrid>
   @override
   void didUpdateWidget(AutoAnimatedSliverGrid oldWidget) {
     super.didUpdateWidget(oldWidget);
-    if (oldWidget.itemCount != widget.itemCount) {
+    if (widget.itemCount < oldWidget.itemCount) {
       init();
     }
   }